Add a pslocal command with good defaults to deploy on PlanktoScope (#236)
* Refactor instrument store methods for linter error about duplicated code

* Try to fix data race in stopping automation job orchestrator

* Don't try to show a blank instrument description

* Allow disabling Ory authn and enabling local admin user

* Add a pslocal command with good defaults to deploy on PlanktoScope RPi

* Run `go mod tidy`

* Fix goreleaser syntax error

* Fix incorrect GOARM for the RPi build

* Fix deprecation warning for goreleaser --rm-dist flag

* Cache modules & builds in Github Actions

* Make separate release archives for pslive and pslocal

* Also cache modules & builds for the release action

* Fix archive name templates

* Also cache modules & builds for the release test action

* Rename goreleaser archives for clarity

* Fix the binary name of the pslocal build

* Upgrade npm dependencies

+++ internal/app/pslive/routes/instruments/instruments.go
@@ -86,6 +86,6 @@ func (h *Handlers) HandleInstrumentsPost() auth.HTTPHandlerFunc {
 		}
 
 		// Redirect user
-		return c.Redirect(http.StatusSeeOther, fmt.Sprintf("/instruments/%d", id))
+		return c.Redirect(http.StatusSeeOther, fmt.Sprintf("/instruments/%d", int64(id)))
 	}
 }
